Structural change for backend implementation
Changed main folder name to avoid conflicts with the PseuToPy package.
