--- pseutopy/blueprints/multilingual/__init__.py
@@ -1,1 +1,0 @@
-from pseutopy.blueprints.multilingual.routes import multilingual
