+++ src/blueprints/multilingual/__init__.py
@@ -1,0 +1,1 @@
+from src.blueprints.multilingual.routes import multilingual
