+++ src/blueprints/multilingual/templates/multilingual/index.html
